requirements.txt
urllib3>=1.10
certifi
python-telegram-bot